Minor Update for Dotenv and Codebase improvements
