+++ Version1.1/AZURE_layer2/app.py
@@ -1,11 +1,10 @@
-from fastapi import FastAPI, Request, HTTPException
+from fastapi import FastAPI
 import logging
 import logging.config
 import signal
 from fastapi.middleware.cors import CORSMiddleware
-from services.langchain_service import generate_positive_suggestion
-from services.sentiment_service import analyze_sentiment
-from services.weaviate_services import store_in_weaviate_with_sentiment
+from AZURE_layer2.routes.echo_route import router as echo_router
+from AZURE_layer2.routes.root_routes import router as root_router
 
 app = FastAPI()
 
@@ -30,55 +29,12 @@ app.add_middleware(
 # Shutdown handler
 def shutdown():
     logging.info("Shutting down...")
-    # Add any cleanup code (e.g., close Weaviate client)
     logging.shutdown()
 
 
 # Register shutdown handler
 signal.signal(signal.SIGINT, lambda sig, frame: shutdown())
 
-# Include your router
-from AZURE_layer2.routes.echo_route import router as echo_router
-
+# Include routers
 app.include_router(echo_router, prefix="/v1/api")
-
-
-@app.get("/")
-async def read_root():
-    return {"message": "Welcome to the API. Use /v1/api for endpoints."}
-
-
-@app.post("/process_echo")
-async def process_echo(request: Request):
-    try:
-        # Parse the incoming JSON payload
-        payload = await request.json()
-        content = payload.get("content")
-        author_id = payload.get("author_id")
-
-        if not content or not author_id:
-            raise HTTPException(status_code=400, detail="Invalid input")
-
-        # Analyze sentiment
-        sentiment = analyze_sentiment(content)
-
-        # Store in Weaviate
-        store_in_weaviate_with_sentiment(content, author_id, sentiment)
-
-        # Generate alternative suggestion if sentiment is neutral or negative
-        alternative = None
-        if sentiment in ["neutral", "negative"]:
-            alternative = generate_positive_suggestion(content)
-
-        # Prepare response
-        response = {
-            "content": content,
-            "author_id": author_id,
-            "sentiment": sentiment,
-            "alternative": alternative,
-        }
-
-        return response
-
-    except Exception as e:
-        raise HTTPException(status_code=500, detail=str(e))
+app.include_router(root_router)
