+++ Version1.1/AZURE_layer2/db/session.py
@@ -1,13 +1,7 @@
-from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
-from dotenv import load_dotenv
-import os
 
-load_dotenv()
-
-DATABASE_URL = os.getenv("DATABASE_URL")
-
-engine = create_engine(DATABASE_URL)
+# Removed DATABASE_URL as Weaviate is the only database being used
+engine = None  # Placeholder for future database engine if needed
 SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)
